Fix issues with inaccurate send rate and limits
Region was permanently set to 'eu-west-1' for development.
However, sending limits for each region are separate
https://docs.aws.amazon.com/ses/latest/DeveloperGuide/manage-sending-limits.html

So when a non-'eu-west-1' region was used, the quotas were reported
as 200 emails/day and 1 emails/second, which is the SES default,
causing large campaigns to be prematurely rejected and sending to
be slow (due to forced rate limiting in the email sending engine).

This was not an obvious issue because there are no authentication
issues, i.e. you can auth with any region, it's just the sending
limits that silently differ.

The region was however passed correctly in the email sending engine,
so forcibly bypassing the send limit (by removing the checking code)
did not cause problems with actual campaign sending, i.e. a campaign
was sent successfully even if num recipients > 200, and it sent at
full speed (not 1 email/second).

> Fix by passing the correct region (from the Settings table) to the
AWS.SES object.