Fix 'from' address SES verification bug
The FROM address was being set to the TO address for each sent email.
This caused SES to reject the email because the sender was not verified.

So emails could only be sent to the owner of the SES account.

Emails sent to the test server were successful because there is
no source address validation.