Cap the max send rate to improve campaign duration estimate accuracy
Part of issue #82

The campaign duration estimate is too optimistic because it assumes we
will be able to send emails at the SES max send rate. Increase the
duration estimate to account for limitations of the app/server so that
we don't underestimate the time required to send a campaign.