Remove default naming
According to the feedback from the meeting held the 31 of May, the lists no longer get a timestamp as a default name. Now the user has to enter a meaningful name before uploading his list. This is done to prevent user to roll with timestamps to get confused down the road when juggling with a few questionably named lists